Move fanOutUserVars to user_variable.go.
