+++ sdk/flow/request.go
@@ -94,43 +94,3 @@ func MarshalChan(replyc chan *grpcflow.Reply, newItem func() proto.Message) (cha
 
 	return itemc, errc
 }
-
-// fanOutUserVars puts user variable setters on their own channel
-func fanOutUserVars(incoming <-chan *grpcflow.Reply, flowsetterc chan<- *grpcflow.Request) (chan *grpcflow.Reply, <-chan *UserVar, chan error) {
-	outgoingc := make(chan *grpcflow.Reply)
-	clientsetterc := make(chan *UserVar)
-	errc := make(chan error)
-
-	go func() {
-		defer func() {
-			close(outgoingc)
-			close(clientsetterc)
-			close(errc)
-			close(flowsetterc)
-		}()
-
-		for msg := range incoming {
-			setter := &grpcflow.UserVariableSetter{}
-			err := ptypes.UnmarshalAny(msg.Payload, setter)
-			if err == nil {
-				// Currently immediately sets the variable to its default value
-				// TODO: pass setter along the chan
-				inner, err := ptypes.MarshalAny(&grpcflow.UserVariableValue{
-					Value: setter.Default,
-					Id:    setter.Id,
-				})
-				if err != nil {
-					errc <- errors.Trace(err)
-				}
-
-				flowsetterc <- &grpcflow.Request{
-					Payload: inner,
-				}
-			} else {
-				outgoingc <- msg
-			}
-		}
-	}()
-
-	return outgoingc, clientsetterc, errc
-}
